requirements.txt
Flask
gunicorn
requests
flask-mysqldb
mysql-connector-python
